+++ game/g_turret.c
@@ -38,9 +38,9 @@ float SnapToEights(float x)
 {
 	x *= 8.0;
 	if (x > 0.0)
-		x += 0.5;
+		x += 0.5f;
 	else
-		x -= 0.5;
+		x -= 0.5f;
 	return 0.125 * (int)x;
 }
 
@@ -116,17 +116,17 @@ void turret_breach_think (edict_t *self)
 	{
 		float	dmin, dmax;
 
-		dmin = fabs(self->pos1[YAW] - self->move_angles[YAW]);
+		dmin = (float)fabs(self->pos1[YAW] - self->move_angles[YAW]);
 		if (dmin < -180)
 			dmin += 360;
 		else if (dmin > 180)
 			dmin -= 360;
-		dmax = fabs(self->pos2[YAW] - self->move_angles[YAW]);
+		dmax = (float)fabs(self->pos2[YAW] - self->move_angles[YAW]);
 		if (dmax < -180)
 			dmax += 360;
 		else if (dmax > 180)
 			dmax -= 360;
-		if (fabs(dmin) < fabs(dmax))
+		if ((float)fabs(dmin) < (float)fabs(dmax))
 			self->move_angles[YAW] = self->pos1[YAW];
 		else
 			self->move_angles[YAW] = self->pos2[YAW];
@@ -180,15 +180,15 @@ void turret_breach_think (edict_t *self)
 		target[2] = self->owner->s.origin[2];
 
 		VectorSubtract (target, self->owner->s.origin, dir);
-		self->owner->velocity[0] = dir[0] * 1.0 / FRAMETIME;
-		self->owner->velocity[1] = dir[1] * 1.0 / FRAMETIME;
+		self->owner->velocity[0] = dir[0] * 1.0f / FRAMETIME;
+		self->owner->velocity[1] = dir[1] * 1.0f / FRAMETIME;
 
 		// z
 		angle = self->s.angles[PITCH] * (M_PI*2 / 360);
 		target_z = SnapToEights(self->s.origin[2] + self->owner->move_origin[0] * tan(angle) + self->owner->move_origin[2]);
 
 		diff = target_z - self->owner->s.origin[2];
-		self->owner->velocity[2] = diff * 1.0 / FRAMETIME;
+		self->owner->velocity[2] = diff * 1.0f / FRAMETIME;
 
 		if (self->spawnflags & 65536)
 		{
@@ -342,11 +342,11 @@ void turret_driver_think (edict_t *self)
 	if (level.time < self->monsterinfo.attack_finished)
 		return;
 
-	reaction_time = (3 - skill->value) * 1.0;
+	reaction_time = (3 - skill->value) * 1.0f;
 	if ((level.time - self->monsterinfo.trail_time) < reaction_time)
 		return;
 
-	self->monsterinfo.attack_finished = level.time + reaction_time + 1.0;
+	self->monsterinfo.attack_finished = level.time + reaction_time + 1.0f;
 	//FIXME how do we really want to pass this along?
 	self->target_ent->spawnflags |= 65536;
 }
